+++ exportplan/helpers.py
@@ -94,3 +94,40 @@ def get_comtrade_historicalimportdata(commodity_code, country):
     response = api_client.dataservices.get_historicalimportdata(commodity_code=commodity_code, country=country)
     response.raise_for_status()
     return response.json()
+
+
+def get_recommended_countries(sso_session_id, sectors):
+    response = api_client.personalisation.recommended_countries_by_sector(sso_session_id=sso_session_id, sector=sectors)
+    response.raise_for_status()
+    parsed = response.json()
+    if parsed:
+        for item in parsed:
+            country = item['country'].title()
+            item['country'] = country
+        return parsed
+    return []
+
+
+def serialize_exportplan_data(rules_regulations, user):
+    target_markets = [{'country': rules_regulations['country']}]
+    if user.company and user.company.expertise_countries_labels:
+        target_markets = target_markets + [{'country': c} for c in user.company.expertise_countries_labels]
+    return {
+        'export_countries': [rules_regulations['country']],
+        'export_commodity_codes': [rules_regulations['commodity_code']],
+        'rules_regulations': rules_regulations,
+        'target_markets': target_markets,
+    }
+
+
+def get_or_create_export_plan(user):
+    # This is a temp hook to create initial export plan. Once we have a full journey this can be removed
+    export_plan = get_exportplan(user.session_id)
+
+    if not export_plan:
+        rules = get_rules_and_regulations('Australia')
+        export_plan = create_export_plan(
+            sso_session_id=user.session_id,
+            exportplan_data=serialize_exportplan_data(rules_regulations=rules, user=user)
+        )
+    return export_plan