+++ exportplan/views.py
@@ -1,19 +1,20 @@
 from datetime import datetime
-import pytz
 import json
 
-from django.views.generic import TemplateView, FormView
+from django.views.generic import TemplateView
 
 from directory_constants.choices import INDUSTRIES
 
-from exportplan import data, forms, helpers
+from exportplan import data, helpers
 
 
 class BaseExportPlanView(TemplateView):
+    export_plan = {}
 
     def get_context_data(self, *args, **kwargs):
         industries = [name for id, name in INDUSTRIES]
         country_choices = [{'value': key, 'label': label} for key, label in helpers.get_madb_country_list()]
+        self.export_plan = helpers.get_or_create_export_plan(self.request.user)
 
         return super().get_context_data(
             next_section=self.next_section,
@@ -49,74 +50,10 @@ class ExportPlanTargetMarketsView(ExportPlanSectionView):
     template_name = 'exportplan/sections/target-markets.html'
 
     def get_context_data(self, *args, **kwargs):
-
-        export_plan = helpers.get_exportplan(sso_session_id=self.request.user.session_id)
-
-        if export_plan:
-            return super().get_context_data(
-                target_markets=export_plan['target_markets'],
-                datenow=datetime.now(),
-                *args, **kwargs)
-
-        return super().get_context_data(*args, *kwargs)
-
-
-class ExportPlanStartView(FormView):
-    template_name = 'exportplan/start.html'
-    form_class = forms.ExportPlanFormStart
-    success_url = '/export-plan/'
-
-    def get_initial(self):
-        return {
-            'country': self.request.GET.get('country'),
-            'commodity': self.request.GET.get('commodity_code'),
-        }
-
-    def get_context_data(self, **kwargs):
-        rules_regulation = None
-        if self.request.GET.get('country'):
-            rules_regulation = helpers.get_rules_and_regulations(self.request.GET['country'])
-        return super().get_context_data(rules_regulation=rules_regulation, **kwargs)
-
-    def form_valid(self, form):
-        rules_regulation = helpers.get_rules_and_regulations(self.request.GET.get('country'))
-        helpers.create_export_plan(
-            sso_session_id=self.request.user.session_id,
-            exportplan_data=self.serialize_exportplan_data(rules_regulation)
-        )
-        return super().form_valid(form)
-
-    def serialize_exportplan_data(self, exportplan_data):
-        return {
-            'export_countries': [exportplan_data['country']],
-            'export_commodity_codes': [exportplan_data['commodity_code']],
-            'rules_regulations': exportplan_data,
-            'target_markets': [{'country': exportplan_data['country']}]
-        }
-
-
-class ExportPlanCreateView(TemplateView):
-    template_name = 'exportplan/create.html'
-
-    def get_context_data(self, **kwargs):
-
-        rules_regulation = helpers.get_exportplan_rules_regulations(sso_session_id=self.request.user.session_id)
-        export_marketdata = helpers.get_exportplan_marketdata(rules_regulation.get('country_code'))
-        utz_offset = datetime.now(pytz.timezone(export_marketdata['timezone'])).strftime('%z')
-        commodity_code = rules_regulation.get('commodity_code')
-        country = rules_regulation.get('country')
-
-        lastyear_import_data = helpers.get_comtrade_lastyearimportdata(commodity_code=commodity_code, country=country)
-        historical_import_data = helpers.get_comtrade_historicalimportdata(
-            commodity_code=commodity_code, country=country
-        )
-
-        return super().get_context_data(
-            rules_regulation=rules_regulation,
-            export_marketdata=export_marketdata,
+        context = super().get_context_data(*args, *kwargs)
+        context.update(
+            selected_sectors=json.dumps(self.export_plan.get('sectors', [])),
+            target_markets=json.dumps(self.export_plan.get('target_markets', [])),
             datenow=datetime.now(),
-            utz_offset=utz_offset,
-            lastyear_import_data=lastyear_import_data,
-            historical_import_data=historical_import_data,
-            **kwargs
         )
+        return context