+++ tests/browser/test_export_plan_about_your_plan.py
@@ -11,7 +11,7 @@ from tests.browser.common_selectors import (
     HeaderSignedIn,
     StickyHeader,
 )
-from tests.browser.util import should_not_see_errors, should_see_all_elements
+from tests.browser.util import should_see_all_elements
 from tests.helpers import create_response
 
 pytestmark = [
@@ -22,6 +22,8 @@ pytestmark = [
 
 
 @pytest.mark.django_db
+@mock.patch.object(exportplan_helpers, 'get_recommended_countries')
+@mock.patch.object(exportplan_helpers, 'update_exportplan')
 @mock.patch.object(exportplan_helpers, 'get_exportplan_marketdata')
 @mock.patch.object(api_client.dataservices, 'get_lastyearimportdata')
 @mock.patch.object(api_client.dataservices, 'get_corruption_perceptions_index')
@@ -39,6 +41,8 @@ def test_export_plan_about_your_business(
     mock_cpi,
     mock_lastyearimportdata,
     mock_get_exportplan_marketdata,
+    mock_update_exportplan,
+    mock_get_recommended_countries,
     server_user_browser_dashboard,
 ):
     mock_update_company_profile.return_value = create_response()
@@ -51,9 +55,13 @@ def test_export_plan_about_your_business(
             'export_commodity_codes': [100],
             'target_markets': [{'country': 'China'}],
             'rules_regulations': {'country_code': 'CHN'},
+            'sectors': ['Automotive'],
+            'pk': 1,
         }
     ]
     mock_get_exportplan.return_value = create_response(data)
+
+    mock_get_recommended_countries.return_value = [{'country': 'Japan'}, {'country': 'South Korea'}]
 
     easeofdoingbusiness_data = {
         'country_name': 'China',
@@ -80,7 +88,6 @@ def test_export_plan_about_your_business(
     mock_get_exportplan_marketdata.return_value = {'timezone': 'Asia/Shanghai', }
 
     live_server, user, browser = server_user_browser_dashboard
-    should_not_see_errors(browser)
 
     browser.get(live_server.url + '/export-plan/section/target-markets/')
     should_see_all_elements(browser, HeaderCommon)